+++ examples/custom-rlpx-subprotocol/src/subprotocol/mod.rs
@@ -1,0 +1,2 @@
+pub(crate) mod connection;
+pub(crate) mod protocol;
