+++ examples/custom-rlpx-subprotocol/src/subprotocol/protocol/event.rs
@@ -1,0 +1,15 @@
+use crate::subprotocol::connection::CustomCommand;
+use reth_network::Direction;
+use reth_network_api::PeerId;
+use tokio::sync::mpsc;
+
+/// The event that can be emitted by our custom protocol.
+#[derive(Debug)]
+pub(crate) enum ProtocolEvent {
+    Established {
+        #[allow(dead_code)]
+        direction: Direction,
+        peer_id: PeerId,
+        to_connection: mpsc::UnboundedSender<CustomCommand>,
+    },
+}
